requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
Django==3.1.7
django-bootstrap4==2.3.1
django-cors-headers==3.10.1
django-environ==0.4.5
djangorestframework==3.13.1
et-xmlfile==1.0.1
idna==3.3
numpy==1.20.2
openpyxl==3.0.7
pandas==1.2.4
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.26.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.7
xlrd==2.0.1
